+++ presto-common/src/main/java/com/facebook/presto/common/array/SliceBigArray.java
@@ -18,8 +18,8 @@ import org.openjdk.jol.info.ClassLayout;
 
 public final class SliceBigArray
 {
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(SliceBigArray.class).instanceSize();
-    private static final int SLICE_INSTANCE_SIZE = ClassLayout.parseClass(Slice.class).instanceSize();
+    private static final int INSTANCE_SIZE = (int) ClassLayout.parseClass(SliceBigArray.class).instanceSize();
+    private static final int SLICE_INSTANCE_SIZE = (int) ClassLayout.parseClass(Slice.class).instanceSize();
     private final ObjectBigArray<Slice> array;
     private final ReferenceCountMap trackedSlices = new ReferenceCountMap();
     private long sizeOfSlices;
