+++ presto-common/src/main/java/com/facebook/presto/common/block/ColumnarArray.java
@@ -20,7 +20,7 @@ import static java.util.Objects.requireNonNull;
 
 public class ColumnarArray
 {
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(ColumnarArray.class).instanceSize();
+    private static final int INSTANCE_SIZE = (int) ClassLayout.parseClass(ColumnarArray.class).instanceSize();
 
     private final Block nullCheckBlock;
     private final int offsetsOffset;
