+++ presto-common/src/main/java/com/facebook/presto/common/block/ColumnarRow.java
@@ -20,7 +20,7 @@ import static java.util.Objects.requireNonNull;
 
 public final class ColumnarRow
 {
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(ColumnarRow.class).instanceSize();
+    private static final int INSTANCE_SIZE = (int) ClassLayout.parseClass(ColumnarRow.class).instanceSize();
 
     private final Block nullCheckBlock;
     private final Block[] fields;
