+++ presto-hive/src/main/java/com/facebook/presto/hive/InternalHiveSplit.java
@@ -38,10 +38,10 @@ import static java.util.Objects.requireNonNull;
 public class InternalHiveSplit
 {
     // Overhead of ImmutableList and ImmutableMap is not accounted for because of its complexity.
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(InternalHiveSplit.class).instanceSize();
+    private static final int INSTANCE_SIZE = (int) ClassLayout.parseClass(InternalHiveSplit.class).instanceSize();
 
-    private static final int HOST_ADDRESS_INSTANCE_SIZE = ClassLayout.parseClass(HostAddress.class).instanceSize() +
-            ClassLayout.parseClass(String.class).instanceSize();
+    private static final int HOST_ADDRESS_INSTANCE_SIZE = (int) ClassLayout.parseClass(HostAddress.class).instanceSize() +
+            (int) ClassLayout.parseClass(String.class).instanceSize();
 
     private final String path;
     private final boolean isRelative;
