+++ presto-main-base/src/main/java/com/facebook/presto/operator/aggregation/NumericHistogram.java
@@ -36,7 +36,7 @@ import static java.util.Objects.requireNonNull;
 public class NumericHistogram
 {
     private static final byte FORMAT_TAG = 0;
-    private static final int INSTANCE_SIZE = ClassLayout.parseClass(NumericHistogram.class).instanceSize();
+    private static final int INSTANCE_SIZE = (int) ClassLayout.parseClass(NumericHistogram.class).instanceSize();
 
     private final int maxBuckets;
     private final double[] values;
